# dist/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Beldex Exchange",
        "description": "This is a core api server endpoints.  You can find out more about Beldex at [https://beldex.io](https://beldex.io).",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "sathish.kumar@beldex.org"
        }
    },
    "host": "coreapialpha.beldex.io",
    "basePath": "/api/v1/",
...
    "schemes": [
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "matching/asset/summary": {
            "get": {
                "tags": [
                    "matching"
                ],
                "description": "Asset Summary",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "asset",
                        "in": "query",
                        "description": "Asset name",
                        "required": false,
                        "type": "String"
                    }
                ]
            }
        },
        "matching/balance/history": {
            "post": {
                "tags": [
                    "matching"
                ],
                "description": "List of balance history",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Parameters",
                        "in": "body",
                        "description": "balance history",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/balanceHistory"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "balance/history",
                        "required": true,
                        "type": "string"
                    }
                ]
            }
        },
        "matching/balance/query": {
            "post": {
                "tags": [
                    "matching"
                ],
                "description": "balance query",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Parameters",
                        "in": "body",
                        "description": "balance query",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/balanceQuery"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "balance/query",
                        "required": true,
                        "type": "string"
                    }
                ]
            }
        },
        "matching/balance/update": {
            "patch": {
                "tags": [
                    "matching"
                ],
                "description": "balance query",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Parameters",
                        "in": "body",
                        "description": "balance update",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/balanceUpdate"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "balance/query",
                        "required": true,
                        "type": "string"
                    }
                ]
            }
        },
        "matching/order/put-market": {
            "post": {
                "tags": [
                    "matching"
                ],
                "description": "Order putMarket",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Parameters",
                        "in": "body",
                        "description": "Order putMarket",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/orderPutMarket"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "order/put-market",
                        "required": true,
                        "type": "string"
                    }
                ]
            }
        },
        "matching/order/put-limit": {
            "post": {
                "tags": [
                    "matching"
                ],
                "description": "Order putLimit",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Parameters",
                        "in": "body",
                        "description": "Order putLimit",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/orderPutLimit"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "order/put-limit",
                        "required": true,
                        "type": "string"
                    }
                ]
            }
        },
        "matching/order/cancel": {
            "post": {
                "tags": [
                    "matching"
                ],
                "description": "Order cancel",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Parameters",
                        "in": "body",
                        "description": "Order Cancel",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/orderCancel"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "order/cancel",
                        "required": true,
                        "type": "string"
                    }
                ]
            }
        },
        "matching/order/book": {
            "post": {
                "tags": [
                    "matching"
                ],
                "description": "Order book",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Parameters",
                        "in": "body",
                        "description": "Order Book",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/orderBook"
                        }
                    }
                    
                ]
            }
        },
        "matching/order/depth": {
            "post": {
                "tags": [
                    "matching"
                ],
                "description": "Order depth",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Parameters",
                        "in": "body",
                        "description": "Order Depth",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/orderDepth"
                        }
                    }
                    
                ]
            }
        },
        "matching/order/pending": {
            "post": {
                "tags": [
                    "matching"
                ],
                "description": "Order pending",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Parameters",
                        "in": "body",
                        "description": "Order Pending",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/orderPending"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "order/pending",
                        "required": true,
                        "type": "string"
                    }
                ]
            }
        },
        "matching/order/pending-detials": {
            "post": {
                "tags": [
                    "matching"
                ],
                "description": "Order pendingDetials",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Parameters",
                        "in": "body",
                        "description": "Order PendingDetials",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/orderPendingDetials"
                        }
                    }                  
                ]
            }
        },
        "matching/order/deals": {
            "post": {
                "tags": [
                    "matching"
                ],
                "description": "Order deals",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Parameters",
                        "in": "body",
                        "description": "Order Deals",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/orderDeals"
                        }
                    }
                   
                ]
            }
        },
        "matching/order/finished": {
            "post": {
                "tags": [
                    "matching"
                ],
                "description": "Order finished",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Parameters",
                        "in": "body",
                        "description": "Order Finished",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/orderFinished"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "order/finsihed",
                        "required": true,
                        "type": "string"
                    }
                ]
            }
        },
        "matching/order/finished-detials": {
            "post": {
                "tags": [
                    "matching"
                ],
                "description": "Order finished-detials",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Parameters",
                        "in": "body",
                        "description": "Order Finished-detials",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/orderFinishedDetials"
                        }
                    }
                ]
            }
        },
        "matching/market/list": {
            "get": {
                "tags": [
                    "matching"
                ],
                "description": "Market list",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                     {
                        "name": "Authorization",
                        "in": "header",
                        "description": "settings",
                        "required": false,
                        "type": "string"
                     }
                ]
            }
        },
        "matching/market/summary": {
            "get": {
                "tags": [
                    "matching"
                ],
                "description": "Market summary",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "asset",
                        "in": "query",
                        "description": "Asset name",
                        "required": false,
                        "type": "String"
                    }
                ]
            }
        },
        "matching/market/status": {
            "post": {
                "tags": [
                    "matching"
                ],
                "description": "Market status",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Parameters",
                        "in": "body",
                        "description": "Market Status",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/marketStatus"
                        }
                    }
                ]
            }
        },
        "matching/market/status-today": {
            "post": {
                "tags": [
                    "matching"
                ],
                "description": "Market statusToday",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Parameters",
                        "in": "body",
                        "description": "Market StatusToday",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/marketStatusToday"
                        }
                    }
                ]
            }
        },
        "matching/market/last": {
            "post": {
                "tags": [
                    "matching"
                ],
                "description": "Market last",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Parameters",
                        "in": "body",
                        "description": "Market Last",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/marketLast"
                        }
                    }
                ]
            }
        },
        "matching/market/deals": {
            "post": {
                "tags": [
                    "matching"
                ],
                "description": "Market deals",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Parameters",
                        "in": "body",
                        "description": "Market Deals",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/marketDeals"
                        }
                    }
                ]
            }
        },
        "matching/market/user-deals": {
            "post": {
                "tags": [
                    "matching"
                ],
                "description": "Market userDeals",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Parameters",
                        "in": "body",
                        "description": "Market UserDeals",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/marketUserDeals"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "market/user-deals",
                        "required": true,
                        "type": "string"
                    }
                ]
            }
        },
        "matching/market/kline": {
            "post": {
                "tags": [
                    "matching"
                ],
                "description": "Market kline",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Parameters",
                        "in": "body",
                        "description": "Market Kline",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/marketKline"
                        }
                    }
                ]
            }
        }
    },
    "definitions": {
...
        "balanceHistory": {
            "properties": {
                "lang": {
                    "required": true,
                    "type": "string",
                    "example": "en"
                },
                "data": {
                    "example": {
                        "attributes": {
                            "asset": "BDX",
                            "business": "ETH ",
                            "start_time": 0,
                            "end_time": 0,
                            "offset": 3,
                            "limit": 2
                        }
                    }
                }
            }
        },
        "balanceQuery": {
            "properties": {
                "lang": {
                    "required": true,
                    "type": "string",
                    "example": "en"
                },
                "data": {
                    "example": {
                        "attributes": {
                            "asset": ["ETH"]
                        }
                    }
                }
            }
        },
        "balanceUpdate": {
            "properties": {
                "lang": {
                    "required": true,
                    "type": "string",
                    "example": "en"
                },
                "data": {
                    "example": {
                        "attributes": {
                            "user_id": 1,
                            "asset": "LTC",
                            "business": "deposit",
                            "business_id": 10,
                            "change": "2000",
                            "detial": {}
                        }
                    }
                }
            }
        },
        "orderPutMarket": {
            "properties": {
                "lang": {
                    "required": true,
                    "type": "string",
                    "example": "en"
                },
                "data": {
                    "example": {
                        "attributes": {
                            "market": "BDXBTC",
                            "side": 1,
                            "amount": "1000",
                            "source": "source added"
                        }
                    }
                }
            }
        },
        "orderPutLimit": {
            "properties": {
                "lang": {
                    "required": true,
                    "type": "string",
                    "example": "en"
                },
                "data": {
                    "example": {
                        "attributes": {
                            "market": "BDXBTC",
                            "side": 1,
                            "amount": "2",
                            "pride": "0.0005",
                            "source": "source added"
                        }
                    }
                }
            }
        },
        "orderCancel": {
            "properties": {
                "lang": {
                    "required": true,
                    "type": "string",
                    "example": "en"
                },
                "data": {
                    "example": {
                        "attributes": {
                            "market": "BDXBTC",
                            "order_id": 2
                        }
                    }
                }
            }
        },
        "orderBook": {
            "properties": {
                "lang": {
                    "required": true,
                    "type": "string",
                    "example": "en"
                },
                "data": {
                    "example": {
                        "attributes": {
                            "market": "BDXBTC",
                            "side": 1,
                            "offset": 0,
                            "limit": 2
                        }
                    }
                }
            }
        },
        "orderDepth": {
            "properties": {
                "lang": {
                    "required": true,
                    "type": "string",
                    "example": "en"
                },
                "data": {
                    "example": {
                        "attributes": {
                            "market": "BDXBTC",
                            "limit": 5,
                            "interval": "1"
                        }
                    }
                }
            }
        },
        "orderPending": {
            "properties": {
                "lang": {
                    "required": true,
                    "type": "string",
                    "example": "en"
                },
                "data": {
                    "example": {
                        "attributes": {
                            "market": "BDXBTC",
                            "offset": 1,
                            "limit": 2
                        }
                    }
                }
            }
        },
        "orderPendingDetials": {
            "properties": {
                "lang": {
                    "required": true,
                    "type": "string",
                    "example": "en"
                },
                "data": {
                    "example": {
                        "attributes": {
                            "market": "BDXBTC",
                            "order_id": 2
                        }
                    }
                }
            }
        },
        "orderDeals": {
            "properties": {
                "lang": {
                    "required": true,
                    "type": "string",
                    "example": "en"
                },
                "data": {
                    "example": {
                        "attributes": {
                            "order_id": 1,
                            "offset": "1",
                            "limit": 3
                        }
                    }
                }
            }
        },
        "orderFinished": {
            "properties": {
                "lang": {
                    "required": true,
                    "type": "string",
                    "example": "en"
                },
                "data": {
                    "example": {
                        "attributes": {
                            "market": "BDXBTC",
                            "start_time": 0,
                            "end_time": 0,
                            "side": 1,
                            "offset": 0,
                            "limit": 1
                        }
                    }
                }
            }
        },
        "orderFinishedDetials": {
            "properties": {
                "lang": {
                    "required": true,
                    "type": "string",
                    "example": "en"
                },
                "data": {
                    "example": {
                        "attributes": {
                            "order_id" :1
                        }
                    }
                }
            }
        },
        "marketStatus": {
            "properties": {
                "lang": {
                    "required": true,
                    "type": "string",
                    "example": "en"
                },
                "data": {
                    "example": {
                        "attributes": {
                            "market":"ETHBTC",
			                "period" : 86400
                        }
                    }
                }
            }
        },
...
    }
}